+++ src/main/java/dev/lambdaurora/lambdynlights/engine/source/DynamicLightSource.java
@@ -11,6 +11,10 @@ package dev.lambdaurora.lambdynlights.engine.source;
 
 import dev.lambdaurora.lambdynlights.engine.lookup.SpatialLookupEntry;
 import it.unimi.dsi.fastutil.longs.LongSet;
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.ChunkSectionPos;
+import net.minecraft.core.Direction;
+import net.minecraft.util.math.MathHelper;
 
 import java.util.stream.Stream;
 
@@ -37,4 +41,40 @@ public interface DynamicLightSource {
 	 * @return the set of chunk sections to rebuild, or an empty set if none is to be rebuilt
 	 */
 	LongSet getDynamicLightChunksToRebuild(boolean forced);
+
+	/**
+	 * Gathers the closest chunks from the given coordinates.
+	 *
+	 * @param chunks the chunk set to add relevant chunks to
+	 * @param x the X-coordinate
+	 * @param y the Y-coordinate
+	 * @param z the Z-coordinate
+	 */
+	static void gatherClosestChunks(LongSet chunks, double x, double y, double z) {
+		var chunkPos = new BlockPos.Mutable(
+				ChunkSectionPos.blockToSectionCoord(x),
+				ChunkSectionPos.blockToSectionCoord(y),
+				ChunkSectionPos.blockToSectionCoord(z)
+		);
+
+		chunks.add(ChunkSectionPos.asLong(chunkPos.getX(), chunkPos.getY(), chunkPos.getZ()));
+
+		var directionX = (MathHelper.floor(x) & 15) >= 8 ? Direction.EAST : Direction.WEST;
+		var directionY = (MathHelper.floor(y) & 15) >= 8 ? Direction.UP : Direction.DOWN;
+		var directionZ = (MathHelper.floor(z) & 15) >= 8 ? Direction.SOUTH : Direction.NORTH;
+
+		for (int i = 0; i < 7; i++) {
+			if (i % 4 == 0) {
+				chunkPos.move(directionX); // X
+			} else if (i % 4 == 1) {
+				chunkPos.move(directionZ); // XZ
+			} else if (i % 4 == 2) {
+				chunkPos.move(directionX.getOpposite()); // Z
+			} else {
+				chunkPos.move(directionZ.getOpposite()); // origin
+				chunkPos.move(directionY); // Y
+			}
+			chunks.add(ChunkSectionPos.asLong(chunkPos.getX(), chunkPos.getY(), chunkPos.getZ()));
+		}
+	}
 }